requirements.txt
# ── Core ─────────────────────────────────────────────
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
pydantic>=2.6,<3.0
pydantic-settings>=2.2,<3.0
websockets>=12.0,<14.0

# ── Data ingestion ───────────────────────────────────
fredapi>=0.5,<1.0
yfinance>=0.2.36,<1.0

# ── Scientific / ML ─────────────────────────────────
numpy>=1.26,<3.0
pandas>=2.2,<3.0
scikit-learn>=1.4,<2.0
hmmlearn>=0.3,<1.0
joblib>=1.3,<2.0
arch>=6.0,<8.0
statsmodels>=0.14,<1.0

# ── Database ─────────────────────────────────────────
psycopg2-binary>=2.9,<3.0

# ── Scheduler ────────────────────────────────────────
APScheduler>=3.10,<4.0

# ── Utilities ────────────────────────────────────────
python-dotenv>=1.0,<2.0
httpx>=0.27,<1.0

# ── AI / Visualisation (Tier-1 enhancements) ────────
anthropic>=0.40,<2.0
plotly>=5.20,<6.0
